extract core functionality
